[FIX] handle testrprogram preconditions in guests precondition list correctly
Virt guests have a list of preconditions and a list of
testprograms. Even thought the usual way should be to have test programs
only in the testprogram part there is no good reason to prevent
testprogram preconditions to show up in the precondition list. This
patch handles such preconditions correctly.