is_msg_valid understands ALL
I added a new virtual state ALL that is only used by is_msg_valid. It
means that a message is valid in all states (Duh!). It removes the need
to name all states e.g. for keep-alive that is always valid. Obviously,
there is no need for a NONE state because what good is a message that is
never valid?