[FIX] path for log file
In live mode we want the log file of the daemonizer (that writes
"app started" and "app stopped") also in /var/log. This path is not
writable during testing. Thus, in testing (e.g. during CPAN install) we
write this log file to /tmp/. This is not perfect because people might
have $TEMP somewhere other than /tmp but we think its good enough.