[FIX] error handling for notify_event
When an error occurs in notify_event we catch it with Try::Tiny. We had
a check on $@ to decide whether to actually print the error
message. This doesn't make sense. First, $@ is not set in the catch part
(this is a feature of Try::Tiny) and when we are in the catch already,
there is no need for a check. The error did occur.