[FIX] install clientpkg only if arch is given
An ssh testrun will install the client package on the target
machine. This is only done when an architecture is given. This way the
user can define whether he wants a (new) client package to be installed
at all. Furthermore, when no arch is given we do not know which client
package is needed so we can not install one. I think it makes more sense
to give a missing arch a new meaning instead of throwing an error.