[FEAT] set config for log_to_file
Installer and PRC messages are usually written to STDOUT/STDERR. This in
turn is captured through console logging. SSH and SimNow preconditions
do not have console logging and thus the output is lost. This patch adds
a new config option that allows Installer and PRC to act accordingly
when we know the console output won't be available.