[RFK] adapt API to WFQ API
Both WFQ and DummyAlgorithm implement function they inherit as stubs
from Algorithm. Dummy is needed as replacement for WFQ in testing
only. Therefore, it should have the same API. This in turn means that
Algorithm also has to match the API of these two.

As a result all algorithms now work on a list of queues received for
get_next_queue. An attribute "queues" and function to manipulate this
attribute are no longer needed.