[TEST] make sure db exists when needed
I reordered the statements for creating the test database from
fixtures. Before the db was not yet ready when it was needed to create
the first objects. I realized this means the "construct_fixture has to
be in a BEGIN block since the DB is needed for "use Controller" which
implies a BEGIN itself.