[TEST] for memory growth
This new unit test checks whether the scheduler growths memory with
objects that the garbage collector can not clean up. This check seems
very system dependend. With to few cycles a memory leak is not detected
with to many cycles it takes extremly long and it looks like the test
detects memory leaks when there are none. I say "it seems" because I
triple checked and did not find one. Still there is a chance I misses
something.