[FEAT] queue denies host
This patch introduces the new feature "queue denies host(s)". It allows
the user to specify a number of hosts that are never choosen for
testruns of that queue. This is implemented by creating a temporary list
of free hosts for that queue that does not contain the forbidden hosts.