[FIX] condition to grep for forbidden hosts
"x ne any(x,y)" does not mean "x does not equal any of (x,y)". It means
"does x not equal any of (x,y)?". Obviously, the former is false (x
equals x) and the later is true (x not equals y). I got this wrong and
this patch fixes it in an nonambigious way.