[FIX] remove deprecated workaround for not starting children simultaniously
In the old days the framework for starting machines (by resetting) was
not able to handle two hosts being resetted at the same time. To prevent
this issue we put a sleep in to make sure the MCP children start with
some delay and thus reach the reset stage at different times which in
turn means they do not reset their machines in a broken way.

Nowadays we have different ways of starting a machine (among them
different resetters) and none of them needs the above mentioned
workaround. Therefore, I removed it.