[RFK] use hostname from testrun_scheduling object where possible
MCP::Child needs a hostname information in many places. Previously, this
information was given as parameter to start_testrun() and subsequent
functions. This does not make sense because the attribute "testrun"
already has this information. Furthermore, the testrun_scheduling
associated with the testrun attribute is given to producers. This means
using the host from the testrun attribute instead of the one given to
the function as argument allows producers to update the hostname without
need to change the producer API.