[FIX] handling attributes in Tapper::MCP object creation
Tapper::MCP classes that have a testrun associated (e.g. MCP::Child)
need the testrun id be set in the new() call. Since the testrun id is
the only required value it makes sense to allow new($id) instead of
new({testrun => $id}). The build function that makes this possible was
broken for the standard way of providing values for attributes (hash
references). It did in fact allow a hash instead of a hash
reference. Since this is nonstandard behaviour I changed it. Now as hash
ref or a single scalar are allowed as parameters for new. The single
scalar is still treated as the testrun id.

The patch also changes a unit test file that used the nonstandard way to
provide multiple attribute values.