[FEAT] put domain specific functions into plugin
Some tasks like console handling or restarting hosts are very different
for different groups using Tapper. Therefore, these tasks should not be
implemented inside the MCP itself but rather by plugins that offer the
needed functionality.

This patch implements the plugin system for this. Plugins need to be
implemented as Moose roles and be named
Tapper::MCP::Plugin::$Task::$Group. A $Task could be "Console" for
example and $Group could be OSRC.

With the patch comes a lot of cleanup. Since console and reboot
functionality is now handled by plugins all console and reboot handling
was removed from the MCP and its tests.

reviewer: schwigon@