make Master poolhost clean
Tapper::MCP::Master keeps track of running Child processes in a hash
that has the test machine name as key. For pool hosts, this breaks
because we now have two processes fighting for the same hash entry and
only the last one wins. This patch changes the hash key to be the
process id of the child process.