[FIX] use testrun id as part of config file
Config files for installer and PRC (created by MCP) used to consist of
hostname and the word "-install" or "-test". This breaks for pools
hosts. A new testrun scheduled on a pool host would overwrite the config
file of other testruns on the same pool host.

With this patch the testrun ID becomes part of the config file
name. This ID is unique and thus the config file is too.