New test_type 'minion'
A test_type 'minion' means to enqueue a job to
a Mojo Minion job queue with the full PRC config.

The job is then picked up by a special minion worker
in the Tapper-PRC side. Both sides handle their states
bi-directional, in particular the finishing via normal
end of program or the stopping of a testrun by sending
the 'stop' command from MCP to Minion, and by PRC to
send the 'test-finished' back to MCP once it cancelled
the corresponding processes.

The MCP Child knows about 'revive' mode and simply does
not re-enqueue the job after an MCP restart.